Fixed beam calculator

+++ src/main/java/app/entities/products/materials/planks/Beam.java
@@ -8,12 +8,12 @@ import java.util.List;
 public class Beam extends Plank {
     private int postGap;
 
-    public Beam(String name, String description, float costPrice, float salesPrice, float width, String unit, List<Integer> preCutsLengths, int height, int postGap) {
+    public Beam(String name, String description, float costPrice, float salesPrice, float width, String unit, List<Integer> preCutsLengths, float height, int postGap) {
         super(name, description, costPrice, salesPrice, width, unit, preCutsLengths, height);
         this.postGap = postGap;
     }
 
-    public Beam(int itemId, String name, String description, float costPrice, float salesPrice, String unit, float width, int height, int postGap) {
+    public Beam(int itemId, String name, String description, float costPrice, float salesPrice, String unit, float width, float height, int postGap) {
         super(itemId, name, description, costPrice, salesPrice, unit, width, height);
         this.postGap = postGap;
     }
@@ -32,7 +32,7 @@ public class Beam extends Plank {
         ps.setString(2, this.getDescription());
         ps.setString(3, this.getUnit());
         ps.setFloat(4, this.getWidth());
-        ps.setInt( 5, this.getHeight());
+        ps.setFloat( 5, this.getHeight());
         ps.setString(6, this.getClass().getSimpleName());
         ps.setNull(7, Types.NUMERIC);
         ps.setInt(8, this.getPostGap());
