+++ src/main/java/app/entities/products/materials/planks/Fascia.java
@@ -10,12 +10,12 @@ public class Fascia extends Plank {
 
     public static int MIN_THICKNESS_FOR_GUTTERS = 22;
 
-    public Fascia(String name, String description, float costPrice, float salesPrice, float width, String unit, List<Integer> preCutsLengths, int height) {
+    public Fascia(String name, String description, float costPrice, float salesPrice, float width, String unit, List<Integer> preCutsLengths, float height) {
         super(name, description, costPrice, salesPrice, width, unit, preCutsLengths, height);
         this.supportsGutters = determineSupportsGutters();
     }
 
-    public Fascia(int itemId, String name, String description, float costPrice, float salesPrice, String unit, float width, int height) {
+    public Fascia(int itemId, String name, String description, float costPrice, float salesPrice, String unit, float width, float height) {
         super(itemId, name, description, costPrice, salesPrice, unit, width, height);
         this.supportsGutters = determineSupportsGutters();
     }
@@ -39,7 +39,7 @@ public class Fascia extends Plank {
         ps.setString(2, this.getDescription());
         ps.setString(3, this.getUnit());
         ps.setFloat(4, this.getWidth());
-        ps.setInt( 5, this.getHeight());
+        ps.setFloat( 5, this.getHeight());
         ps.setString(6, this.getClass().getSimpleName());
         ps.setNull(7, Types.NUMERIC);
         ps.setNull(8, Types.NUMERIC);
